requirements.txt
fastapi
uvicorn
httpx
numpy
python-multipart
scikit-learn
python-dotenv
mistralai
